Fix merging of raw BERT models
