+++ mergekit/_data/architectures/bert.json
@@ -5,29 +5,29 @@
     ],
     "pre_weights": [
         {
-            "name": "bert.embeddings.position_embeddings.weight"
+            "name": "embeddings.position_embeddings.weight"
         },
         {
-            "name": "bert.embeddings.token_type_embeddings.weight"
+            "name": "embeddings.token_type_embeddings.weight"
         },
         {
-            "name": "bert.embeddings.word_embeddings.weight",
+            "name": "embeddings.word_embeddings.weight",
             "is_embed": true
         },
         {
-            "name": "bert.embeddings.LayerNorm.bias",
+            "name": "embeddings.LayerNorm.bias",
             "aliases": [
-                "bert.embeddings.LayerNorm.beta"
+                "embeddings.LayerNorm.beta"
             ]
         },
         {
-            "name": "bert.embeddings.LayerNorm.weight",
+            "name": "embeddings.LayerNorm.weight",
             "aliases": [
-                "bert.embeddings.LayerNorm.gamma"
+                "embeddings.LayerNorm.gamma"
             ]
         },
         {
-            "name": "bert.embeddings.position_ids",
+            "name": "embeddings.position_ids",
             "optional": true,
             "force_dtype": "int64"
         }
@@ -44,63 +44,63 @@
     "layer_templates": {
         "weights": [
             {
-                "name": "bert.encoder.layer.${layer_index}.attention.self.query.weight"
+                "name": "encoder.layer.${layer_index}.attention.self.query.weight"
             },
             {
-                "name": "bert.encoder.layer.${layer_index}.attention.self.query.bias"
+                "name": "encoder.layer.${layer_index}.attention.self.query.bias"
             },
             {
-                "name": "bert.encoder.layer.${layer_index}.attention.self.key.weight"
+                "name": "encoder.layer.${layer_index}.attention.self.key.weight"
             },
             {
-                "name": "bert.encoder.layer.${layer_index}.attention.self.key.bias"
+                "name": "encoder.layer.${layer_index}.attention.self.key.bias"
             },
             {
-                "name": "bert.encoder.layer.${layer_index}.attention.self.value.weight"
+                "name": "encoder.layer.${layer_index}.attention.self.value.weight"
             },
             {
-                "name": "bert.encoder.layer.${layer_index}.attention.self.value.bias"
+                "name": "encoder.layer.${layer_index}.attention.self.value.bias"
             },
             {
-                "name": "bert.encoder.layer.${layer_index}.attention.output.dense.weight"
+                "name": "encoder.layer.${layer_index}.attention.output.dense.weight"
             },
             {
-                "name": "bert.encoder.layer.${layer_index}.attention.output.dense.bias"
+                "name": "encoder.layer.${layer_index}.attention.output.dense.bias"
             },
             {
-                "name": "bert.encoder.layer.${layer_index}.attention.output.LayerNorm.bias",
+                "name": "encoder.layer.${layer_index}.attention.output.LayerNorm.bias",
                 "aliases": [
-                    "bert.encoder.layer.${layer_index}.attention.output.LayerNorm.beta"
+                    "encoder.layer.${layer_index}.attention.output.LayerNorm.beta"
                 ]
             },
             {
-                "name": "bert.encoder.layer.${layer_index}.attention.output.LayerNorm.weight",
+                "name": "encoder.layer.${layer_index}.attention.output.LayerNorm.weight",
                 "aliases": [
-                    "bert.encoder.layer.${layer_index}.attention.output.LayerNorm.gamma"
+                    "encoder.layer.${layer_index}.attention.output.LayerNorm.gamma"
                 ]
             },
             {
-                "name": "bert.encoder.layer.${layer_index}.intermediate.dense.weight"
+                "name": "encoder.layer.${layer_index}.intermediate.dense.weight"
             },
             {
-                "name": "bert.encoder.layer.${layer_index}.intermediate.dense.bias"
+                "name": "encoder.layer.${layer_index}.intermediate.dense.bias"
             },
             {
-                "name": "bert.encoder.layer.${layer_index}.output.dense.weight"
+                "name": "encoder.layer.${layer_index}.output.dense.weight"
             },
             {
-                "name": "bert.encoder.layer.${layer_index}.output.dense.bias"
+                "name": "encoder.layer.${layer_index}.output.dense.bias"
             },
             {
-                "name": "bert.encoder.layer.${layer_index}.output.LayerNorm.bias",
+                "name": "encoder.layer.${layer_index}.output.LayerNorm.bias",
                 "aliases": [
-                    "bert.encoder.layer.${layer_index}.output.LayerNorm.beta"
+                    "encoder.layer.${layer_index}.output.LayerNorm.beta"
                 ]
             },
             {
-                "name": "bert.encoder.layer.${layer_index}.output.LayerNorm.weight",
+                "name": "encoder.layer.${layer_index}.output.LayerNorm.weight",
                 "aliases": [
-                    "bert.encoder.layer.${layer_index}.output.LayerNorm.gamma"
+                    "encoder.layer.${layer_index}.output.LayerNorm.gamma"
                 ]
             }
         ]
