+++ mergekit/common.py
@@ -23,7 +23,6 @@ from typing import (
     Dict,
     Generic,
     Iterator,
-    List,
     Mapping,
     Optional,
     Tuple,
@@ -207,33 +206,6 @@ def dtype_from_name(name: Optional[str]) -> Optional[torch.dtype]:
     elif name == "int64":
         return torch.int64
     raise RuntimeError(f'Unimplemented dtype "{name}"')
-
-
-def rectify_embed_sizes(param_name: str, tensors: List[torch.Tensor]):
-    # TODO: use arch_info.embed_weights() instead
-    if ("lm_head" in param_name or "embed_tokens" in param_name) and all(
-        len(t.shape) == 2 for t in tensors
-    ):
-        # special case - if lm_head.weight or embed_tokens.weight have a size
-        # mismatch, take the largest common submatrix of all of them
-        if take_common_submatrix(tensors):
-            logging.warning(
-                f"Using common submatrix of size {tensors[0].shape} for {param_name}"
-            )
-
-
-def take_common_submatrix(tensors: List[torch.Tensor]) -> bool:
-    min_size = [None, None]
-    for t in tensors:
-        for idx in range(2):
-            if min_size[idx] is None or t.shape[idx] < min_size[idx]:
-                min_size[idx] = t.shape[idx]
-
-    if not all(t.shape == torch.Size(min_size) for t in tensors):
-        for idx in range(len(tensors)):
-            tensors[idx] = tensors[idx][: min_size[0], : min_size[1]]
-        return True
-    return False
 
 
 def parse_kmb(value: Union[str, int]) -> int:
