+++ mergekit/merge_methods/rectify_embed.py
@@ -1,0 +1,47 @@
+# Copyright (C) 2024 Charles O. Goddard
+#
+# This software is free software: you can redistribute it and/or
+# modify it under the terms of the GNU Lesser General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This software is distributed in the hope that it will be useful, but
+# WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+# Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public License
+# along with this program. If not, see http://www.gnu.org/licenses/.
+
+
+import logging
+from typing import List
+
+import torch
+
+from mergekit.architecture import WeightInfo
+
+
+def rectify_embed_sizes(weight_info: WeightInfo, tensors: List[torch.Tensor]):
+    # TODO: use arch_info.embed_weights() instead
+    if weight_info.is_embed and all(len(t.shape) == 2 for t in tensors):
+        # special case - if lm_head.weight or embed_tokens.weight have a size
+        # mismatch, take the largest common submatrix of all of them
+        if take_common_submatrix(tensors):
+            logging.warning(
+                f"Using common submatrix of size {tensors[0].shape} for {weight_info.name}"
+            )
+
+
+def take_common_submatrix(tensors: List[torch.Tensor]) -> bool:
+    min_size = [None, None]
+    for t in tensors:
+        for idx in range(2):
+            if min_size[idx] is None or t.shape[idx] < min_size[idx]:
+                min_size[idx] = t.shape[idx]
+
+    if not all(t.shape == torch.Size(min_size) for t in tensors):
+        for idx in range(len(tensors)):
+            tensors[idx] = tensors[idx][: min_size[0], : min_size[1]]
+        return True
+    return False
